warehouse-preposition-optimizer/pyproject.toml
[project]
name = "warehouse-preposition-optimizer"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "redis>=5.0",
    "httpx>=0.27",
    "structlog>=24.0",
    "prometheus-client>=0.21",
    "python-multipart>=0.0.12",
    "pyyaml>=6.0",
    # Phase 2: ML demand prediction
    "lightgbm>=4.0",
    "shap>=0.46",
    "optuna>=3.0",
    "scikit-learn>=1.5",
    "pandas>=2.0",
    "numpy>=1.26",
    "ortools>=9.15.6755",
    "simpy>=4.1.1",
    "gymnasium>=1.2.3",
    # Floor plan import
    "ezdxf>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "ruff>=0.8",
    "mypy>=1.13",
    "httpx>=0.27",
    "fakeredis[aioredis]>=2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
select = ["I", "E", "W", "F"]

[tool.mypy]
strict = true
python_version = "3.12"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[dependency-groups]
dev = [
    "numpy>=2.4.4",
]
